tools/Video Generator/requirements.txt
# Core Dependencies
requests>=2.31.0
python-dotenv>=1.0.0
Pillow>=10.0.0

# Replicate Dependencies
replicate>=0.25.0

# Google Gemini Dependencies (Veo)
google-genai>=1.0.0

# Local Video Editing
moviepy>=2.0.0
imageio-ffmpeg>=0.4.9

